+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/EdcSubmodelClient.java
@@ -36,15 +36,17 @@ import io.github.resilience4j.retry.RetryRegistry;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.validator.routines.UrlValidator;
+import org.eclipse.dataspaceconnector.spi.types.domain.catalog.Catalog;
 import org.eclipse.dataspaceconnector.spi.types.domain.edr.EndpointDataReference;
-import org.eclipse.tractusx.irs.edc.client.exceptions.EdcClientException;
-import org.eclipse.tractusx.irs.edc.client.model.NegotiationResponse;
-import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotification;
-import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotificationResponse;
 import org.eclipse.tractusx.irs.common.CxTestDataContainer;
 import org.eclipse.tractusx.irs.common.Masker;
 import org.eclipse.tractusx.irs.common.OutboundMeterRegistryService;
 import org.eclipse.tractusx.irs.component.Relationship;
+import org.eclipse.tractusx.irs.edc.client.exceptions.EdcClientException;
+import org.eclipse.tractusx.irs.edc.client.model.CatalogItem;
+import org.eclipse.tractusx.irs.edc.client.model.NegotiationResponse;
+import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotification;
+import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotificationResponse;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StopWatch;
@@ -61,6 +63,9 @@ public interface EdcSubmodelClient {
 
     CompletableFuture<EdcNotificationResponse> sendNotification(String submodelEndpointAddress, String assetId,
             EdcNotification notification) throws EdcClientException;
+
+    CompletableFuture<EndpointDataReference> getEndpointReferenceForAsset(String endpointAddress, String filterKey,
+            String filterValue) throws EdcClientException;
 }
 
 /**
@@ -102,6 +107,12 @@ class EdcSubmodelClientLocalStub implements EdcSubmodelClient {
         // not actually sending anything, just return success response
         return CompletableFuture.completedFuture(() -> true);
     }
+
+    @Override
+    public CompletableFuture<EndpointDataReference> getEndpointReferenceForAsset(final String endpointAddress,
+            final String filterKey, final String filterValue) throws EdcClientException {
+        throw new EdcClientException("Not implemented");
+    }
 }
 
 /**
@@ -111,6 +122,7 @@ class EdcSubmodelClientLocalStub implements EdcSubmodelClient {
 @Slf4j
 @RequiredArgsConstructor
 @Profile({ "!local && !stubtest" })
+@SuppressWarnings("PMD.TooManyMethods")
 class EdcSubmodelClientImpl implements EdcSubmodelClient {
 
     private final EdcConfiguration config;
@@ -120,7 +132,14 @@ class EdcSubmodelClientImpl implements EdcSubmodelClient {
     private final AsyncPollingService pollingService;
     private final OutboundMeterRegistryService meterRegistryService;
     private final RetryRegistry retryRegistry;
+    private final CatalogCache catalogCache;
+    private final EdcControlPlaneClient edcControlPlaneClient;
     private final UrlValidator urlValidator = new UrlValidator(UrlValidator.ALLOW_LOCAL_URLS);
+
+    private static void stopWatchOnEdcTask(final StopWatch stopWatch) {
+        stopWatch.stop();
+        log.info("EDC Task '{}' took {} ms", stopWatch.getLastTaskName(), stopWatch.getLastTaskTimeMillis());
+    }
 
     @Override
     public CompletableFuture<List<Relationship>> getRelationships(final String submodelEndpointAddress,
@@ -149,9 +168,12 @@ class EdcSubmodelClientImpl implements EdcSubmodelClient {
         final String providerConnectorUrl = submodelEndpointAddress.substring(0, indexOfUrn);
         final String target = submodelEndpointAddress.substring(indexOfUrn + 1, indexOfSubModel);
         final String decodedTarget = URLDecoder.decode(target, StandardCharsets.UTF_8);
-        log.info("Starting contract negotiation with providerConnectorUrl {} and target {}", providerConnectorUrl,
+        final String providerWithSuffix = appendSuffix(providerConnectorUrl,
+                config.getControlplane().getProviderSuffix());
+        log.info("Starting contract negotiation with providerConnectorUrl {} and target {}", providerWithSuffix,
                 decodedTarget);
-        return contractNegotiationService.negotiate(providerConnectorUrl, decodedTarget);
+        final CatalogItem catalogItem = catalogCache.getCatalogItem(providerWithSuffix, decodedTarget).orElseThrow();
+        return contractNegotiationService.negotiate(providerWithSuffix, catalogItem);
     }
 
     private CompletableFuture<List<Relationship>> startSubmodelDataRetrieval(
@@ -191,32 +213,42 @@ class EdcSubmodelClientImpl implements EdcSubmodelClient {
 
     private Optional<String> retrieveSubmodelData(final String submodel, final String contractAgreementId,
             final StopWatch stopWatch) {
-        log.info("Retrieving dataReference from storage for contractAgreementId {}", Masker.mask(contractAgreementId));
-        final Optional<EndpointDataReference> dataReference = endpointDataReferenceStorage.remove(contractAgreementId);
+        final Optional<EndpointDataReference> dataReference = retrieveEndpointDataReference(contractAgreementId);
 
         if (dataReference.isPresent()) {
             final EndpointDataReference ref = dataReference.get();
             log.info("Retrieving data from EDC data plane for dataReference with id {}", ref.getId());
             final String data = edcDataPlaneClient.getData(ref, submodel);
-            stopWatch.stop();
-            log.info("EDC Task '{}' took {} ms", stopWatch.getLastTaskName(), stopWatch.getLastTaskTimeMillis());
+            stopWatchOnEdcTask(stopWatch);
 
             return Optional.of(data);
         }
         return Optional.empty();
     }
 
+    private Optional<EndpointDataReference> retrieveEndpointReference(final String contractAgreementId,
+            final StopWatch stopWatch) {
+        final Optional<EndpointDataReference> dataReference = retrieveEndpointDataReference(contractAgreementId);
+
+        if (dataReference.isPresent()) {
+            final EndpointDataReference ref = dataReference.get();
+            log.info("Retrieving Endpoint Reference data from EDC data plane with id: {}", ref.getId());
+            stopWatchOnEdcTask(stopWatch);
+
+            return Optional.of(ref);
+        }
+        return Optional.empty();
+    }
+
     private Optional<EdcNotificationResponse> sendSubmodelNotification(final String submodel,
             final String contractAgreementId, final EdcNotification notification, final StopWatch stopWatch) {
-        log.info("Retrieving dataReference from storage for contractAgreementId {}", contractAgreementId);
-        final Optional<EndpointDataReference> dataReference = endpointDataReferenceStorage.remove(contractAgreementId);
+        final Optional<EndpointDataReference> dataReference = retrieveEndpointDataReference(contractAgreementId);
 
         if (dataReference.isPresent()) {
             final EndpointDataReference ref = dataReference.get();
             log.info("Sending data to EDC data plane with dataReference {}:{}", ref.getAuthKey(), ref.getAuthCode());
             final EdcNotificationResponse response = edcDataPlaneClient.sendData(ref, submodel, notification);
-            stopWatch.stop();
-            log.info("EDC Task '{}' took {} ms", stopWatch.getLastTaskName(), stopWatch.getLastTaskTimeMillis());
+            stopWatchOnEdcTask(stopWatch);
             return Optional.of(response);
         }
         return Optional.empty();
@@ -255,6 +287,55 @@ class EdcSubmodelClientImpl implements EdcSubmodelClient {
 
             return sendNotificationAsync(negotiationResponse.getContractAgreementId(), notification, stopWatch);
         });
+    }
+
+    @Override
+    public CompletableFuture<EndpointDataReference> getEndpointReferenceForAsset(final String endpointAddress,
+            final String filterKey, final String filterValue) throws EdcClientException {
+        return execute(endpointAddress, () -> {
+            final StopWatch stopWatch = new StopWatch();
+            stopWatch.start("Get EDC Submodel task for shell descriptor, endpoint " + endpointAddress);
+            final String providerWithSuffix = appendSuffix(endpointAddress, config.getControlplane().getProviderSuffix());
+
+            final Catalog catalog = edcControlPlaneClient.getCatalogWithFilter(providerWithSuffix, filterKey, filterValue);
+
+            final List<CatalogItem> items = catalog.getContractOffers()
+                                                   .stream()
+                                                   .map(contractOffer -> CatalogItem.builder()
+                                                                                    .itemId(contractOffer.getId())
+                                                                                    .assetPropId(
+                                                                                            contractOffer.getAsset()
+                                                                                                         .getId())
+                                                                                    .connectorId(catalog.getId())
+                                                                                    .policy(contractOffer.getPolicy())
+                                                                                    .build())
+                                                   .toList();
+            final NegotiationResponse response = contractNegotiationService.negotiate(providerWithSuffix,
+                    items.stream().findFirst().orElseThrow());
+
+            return pollingService.<EndpointDataReference>createJob()
+                                 .action(() -> retrieveEndpointReference(response.getContractAgreementId(), stopWatch))
+                                 .timeToLive(config.getSubmodel().getRequestTtl())
+                                 .description("waiting for Endpoint Reference retrieval")
+                                 .build()
+                                 .schedule();
+
+        });
+    }
+
+    private String appendSuffix(final String endpointAddress, final String providerSuffix) {
+        String addressWithSuffix;
+        if (endpointAddress.endsWith("/") && providerSuffix.startsWith("/")) {
+            addressWithSuffix = endpointAddress.substring(0, endpointAddress.length() - 1) + providerSuffix;
+        } else {
+            addressWithSuffix = endpointAddress + providerSuffix;
+        }
+        return addressWithSuffix;
+    }
+
+    private Optional<EndpointDataReference> retrieveEndpointDataReference(final String contractAgreementId) {
+        log.info("Retrieving dataReference from storage for contractAgreementId {}", Masker.mask(contractAgreementId));
+        return endpointDataReferenceStorage.remove(contractAgreementId);
     }
 
     @SuppressWarnings({ "PMD.AvoidRethrowingException",